Refine wallet deposit processing

+++ src/Services/WalletService/Licit.WalletService.Application/Features/CQRS/Wallet/Commands/Deposit/DepositFundsCommandRequest.cs
@@ -4,7 +4,7 @@ namespace Licit.WalletService.Application.Features.CQRS.Wallet.Commands.Deposit;
 
 public record DepositFundsCommandRequest(
     Guid UserId,
-    decimal Amount,
+    int Amount,
     string IdempotencyKey,
     Guid? ReferenceId = null,
     string? Description = null
